// src/types/commands.ts
export enum Commands {
    DisplayTimer = 'pomotimer.displayTimer',
    StartTimer = 'pomotimer.startTimer',
    PauseTimer = 'pomotimer.pauseTimer',
    ResumeTimer = 'pomotimer.resumeTimer',
    RestartTimer = 'pomotimer.restartTimer',
    HideTimer = 'pomotimer.hideTimer',
    DisplayTaskboard = 'pomotimer.displayTaskboard'
}